+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.18.0
 * Improve docs for isPaired
+* Add Streams: `scanStream`, `availabilityStream`, `connectionStream`, `characteristicValueStream`, `pairingStateStream`
 
 ## 0.17.0
 * Fix Windows crash when calling pair APIs with an unknown deviceId